4.21

+++ Part4/blog_list_app_backend/controllers/blogs.js
@@ -45,8 +45,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
+    const userid = user.id
+
+    if (blog.user.toString() === userid.toString()) {
+
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'User is not allowed to delete this post' })
+    }
 })
 
 /* changed this to the async/await
